requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
openai
python-dotenv
fastapi
langserve
sse_starlette